float balance

+++ interface.go
@@ -163,17 +163,17 @@ type CpsMessages struct {
 
 type CompetitionDataResponses struct {
 	Data []struct {
-		ContestID    string `json:"contest_id"`
-		CustomerID   int    `json:"customer_id"`
-		FxID         string `json:"fx_id"`
-		FxMasterPw   string `json:"fx_master_pw"`
-		FxInvesterPw string `json:"fx_invester_pw"`
-		StatusID     int    `json:"status_id"`
-		Balance      int    `json:"balance"`
-		Equity       int    `json:"equity"`
-		Profit       int    `json:"profit"`
-		StartBalance int    `json:"start_balance"`
-		BuyinStatus  bool   `json:"buyin_status"`
+		ContestID    string  `json:"contest_id"`
+		CustomerID   int     `json:"customer_id"`
+		FxID         string  `json:"fx_id"`
+		FxMasterPw   string  `json:"fx_master_pw"`
+		FxInvesterPw string  `json:"fx_invester_pw"`
+		StatusID     float64 `json:"status_id"`
+		Balance      float64 `json:"balance"`
+		Equity       float64 `json:"equity"`
+		Profit       float64 `json:"profit"`
+		StartBalance float64 `json:"start_balance"`
+		BuyinStatus  bool    `json:"buyin_status"`
 	} `json:"data"`
 }
